+++ src/Contacts/src/ContactDetector.cpp
@@ -6,10 +6,11 @@
  */
 
 #include <BipedalLocomotion/ContactDetectors/ContactDetector.h>
+
 using namespace BipedalLocomotion::ParametersHandler;
 using namespace BipedalLocomotion::Contacts;
 
-bool ContactDetector::initialize(std::weak_ptr<IParametersHandler> handler)
+bool ContactDetector::initialize(std::weak_ptr<const IParametersHandler> handler)
 {
     std::string_view printPrefix = "[ContactDetector::initialize] ";
     if (m_detectorState != State::NotInitialized)
@@ -44,7 +45,7 @@ bool ContactDetector::initialize(std::weak_ptr<IParametersHandler> handler)
     return true;
 }
 
-bool ContactDetector::customInitialization(std::weak_ptr<IParametersHandler> handler)
+bool ContactDetector::customInitialization(std::weak_ptr<const IParametersHandler> handler)
 {
     return true;
 }
@@ -86,7 +87,7 @@ bool ContactDetector::resetContacts()
 }
 
 
-const EstimatedContactList& ContactDetector::get() const
+const EstimatedContactList& ContactDetector::getOutput() const
 {
     return m_contactStates;
 }
@@ -114,9 +115,7 @@ EstimatedContact ContactDetector::get(const std::string& contactName) const
     return m_contactStates.at(contactName);
 }
 
-bool ContactDetector::isValid() const
+bool ContactDetector::isOutputValid() const
 {
     return (m_detectorState == State::Running);
 }
-
-