+++ src/Perception/src/ArucoDetector.cpp
@@ -1,13 +1,13 @@
 /**
  * @file ArucoDetector.cpp
  * @authors Prashanth Ramadoss
- * @copyright 2020 Istituto Italiano di Tecnologia (IIT). This software may be modified and
+ * @copyright 2021 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
  */
 
-#include <BipedalLocomotion/Perception/Features/ArucoDetector.h>
-#include <BipedalLocomotion/GenericContainer/Vector.h>
 #include <BipedalLocomotion/Conversions/CommonConversions.h>
+#include <BipedalLocomotion/GenericContainer/Vector.h>
+#include <BipedalLocomotion/Perception/Features/ArucoDetector.h>
 #include <BipedalLocomotion/TextLogging/Logger.h>
 
 using namespace BipedalLocomotion::ParametersHandler;
@@ -29,16 +29,21 @@ public:
 
     // parameters
     cv::Ptr<cv::aruco::Dictionary> dictionary; /**< container with detected markers data */
-    double markerLength;  /**< marker length*/
+    double markerLength; /**< marker length*/
     cv::Mat cameraMatrix; /**< camera calibration matrix*/
-    cv::Mat distCoeff;    /**< camera distortion coefficients*/
+    cv::Mat distCoeff; /**< camera distortion coefficients*/
 
     ArucoDetectorOutput out; /**< container with detected markers data */
     cv::Mat currentImg; /**< currently set image */
     double currentTime{-1.0}; /** time at which currentImg was set */
     std::vector<int> currentDetectedMarkerIds; /**< currently detected marker ids */
-    std::vector<std::vector<cv::Point2f> > currentDetectedMarkerCorners; /**< currently detected marker corners */
-    std::vector<cv::Vec3d> currentDetectedMarkersRotVecs, currentDetectedMarkersTransVecs; /**< currently detected marker rotation and translation vectors */
+    std::vector<std::vector<cv::Point2f>> currentDetectedMarkerCorners; /**< currently detected
+                                                                           marker corners */
+
+    std::vector<cv::Vec3d> currentDetectedMarkersRotVecs; /**< currently detected marker
+                                                               rotation vector */
+    std::vector<cv::Vec3d> currentDetectedMarkersTransVecs; /**< currently detected marker
+                                                               translation vector */
 
     bool initialized{false}; /**< flag to check if detector was initialized properly */
 
@@ -50,63 +55,63 @@ public:
     /**
      * Utility map for choosing Aruco marker dictionary depending on user parameter
      */
-    std::unordered_map<std::string, cv::aruco::PREDEFINED_DICTIONARY_NAME> availableDict{{"4X4_50", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_50},
-                                                                                         {"4X4_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_100},
-                                                                                         {"4X4_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_250},
-                                                                                         {"4X4_1000",cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_1000},
-                                                                                         {"5X5_50",  cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_50},
-                                                                                         {"5X5_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_100},
-                                                                                         {"5X5_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_250},
-                                                                                         {"5X5_1000",cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_1000},
-                                                                                         {"6X6_50",  cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_50},
-                                                                                         {"6X6_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_100},
-                                                                                         {"6X6_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_250},
-                                                                                         {"6X6_1000",cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_1000},
-                                                                                         {"7X7_50",  cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_50},
-                                                                                         {"7X7_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_100},
-                                                                                         {"7X7_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_250},
-                                                                                         {"7X7_1000",cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_1000},
-                                                                                         {"ARUCO_ORIGINAL",cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_ARUCO_ORIGINAL} };
+    std::unordered_map<std::string, cv::aruco::PREDEFINED_DICTIONARY_NAME>
+        availableDict{{"4X4_50", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_50},
+                      {"4X4_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_100},
+                      {"4X4_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_250},
+                      {"4X4_1000", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_4X4_1000},
+                      {"5X5_50", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_50},
+                      {"5X5_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_100},
+                      {"5X5_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_250},
+                      {"5X5_1000", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_5X5_1000},
+                      {"6X6_50", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_50},
+                      {"6X6_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_100},
+                      {"6X6_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_250},
+                      {"6X6_1000", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_6X6_1000},
+                      {"7X7_50", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_50},
+                      {"7X7_100", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_100},
+                      {"7X7_250", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_250},
+                      {"7X7_1000", cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_7X7_1000},
+                      {"ARUCO_ORIGINAL",
+                       cv::aruco::PREDEFINED_DICTIONARY_NAME::DICT_ARUCO_ORIGINAL}};
 };
 
-ArucoDetector::ArucoDetector() :
-               m_pimpl(std::make_unique<Impl>())
+ArucoDetector::ArucoDetector()
+    : m_pimpl(std::make_unique<Impl>())
 {
 }
 
 ArucoDetector::~ArucoDetector() = default;
 
-bool ArucoDetector::initialize(std::weak_ptr<IParametersHandler> handler)
+bool ArucoDetector::initialize(std::weak_ptr<const IParametersHandler> handler)
 {
-    std::string_view printPrefix = "[ArucoDetector::initialize] ";
+    constexpr auto printPrefix = "[ArucoDetector::initialize]";
     auto handle = handler.lock();
     if (handle == nullptr)
     {
-        log()->error("{} "
-        "The parameter handler has expired. Please check its scope.", printPrefix);
+        log()->error("{} The parameter handler has expired. Please check its scope.", printPrefix);
         return false;
     }
 
     std::string dictName;
     if (!handle->getParameter("marker_dictionary", dictName))
     {
-        log()->error("{} "
-        "The parameter handler could not find \" marker_dictionary \" in the configuration file.",
-        printPrefix);
+        log()->error("{} The parameter handler could not find \" marker_dictionary \" in the "
+                     "configuration file.",
+                     printPrefix);
         return false;
     }
 
     if (m_pimpl->availableDict.find(dictName) == m_pimpl->availableDict.end())
     {
-        log()->error( "{} "
-        "Undefined value set to \" marker_dictionary \" in the configuration file. Please choose one among \n"
-        "available options: \"4X4_50\", \"4X4_100\", \"4X4_250\", \"4X4_1000\", \n"
-        "\"5X5_50\", \"5X5_100\", \"5X5_250\", \"5X5_1000\", \n"
-        "\"6X6_50\", \"6X6_100\", \"6X6_250\", \"6X6_1000\", \n"
-        "\"7X7_50\", \"7X7_100\", \"7X7_250\", \"7X7_1000\", \n"
-        " \"ARUCO_ORIGINAL\", \n"
-        "options coherent with v3.4.0 in https://docs.opencv.org/3.4.1/d9/d6a/group__aruco.html",
-        printPrefix);
+        log()->error("{} Undefined value set to \" marker_dictionary \" in the configuration file. "
+                     "Please choose one among \n available options: \"4X4_50\", \"4X4_100\", "
+                     "\"4X4_250\", \"4X4_1000\", \n \"5X5_50\", \"5X5_100\", \"5X5_250\", "
+                     "\"5X5_1000\", \n \"6X6_50\", \"6X6_100\", \"6X6_250\", \"6X6_1000\", \n "
+                     "\"7X7_50\", \"7X7_100\", \"7X7_250\", \"7X7_1000\", \n \"ARUCO_ORIGINAL\", "
+                     "\n options coherent with v3.4.0 in "
+                     "https://docs.opencv.org/3.4.1/d9/d6a/group__aruco.html",
+                     printPrefix);
         return false;
     }
 
@@ -114,31 +119,32 @@ bool ArucoDetector::initialize(std::weak_ptr<IParametersHandler> handler)
 
     if (!handle->getParameter("marker_length", m_pimpl->markerLength))
     {
-        log()->error("{} "
-        "The parameter handler could not find \" marker_length \" in the configuration file.",
-        printPrefix);
+        log()->error("{} The parameter handler could not find \" marker_length \" in the "
+                     "configuration file.",
+                     printPrefix);
         return false;
     }
 
     Eigen::Matrix<double, 9, 1> calibVec;
     if (!handle->getParameter("camera_matrix", calibVec))
     {
-        log()->error("{} "
-        "The parameter handler could not find \" camera_matrix \" in the configuration file.",
-        printPrefix);
+        log()->error("{} The parameter handler could not find \" camera_matrix \" in the "
+                     "configuration file.",
+                     printPrefix);
         return false;
     }
 
-    Eigen::Matrix3d calibMat = Eigen::Map< Eigen::Matrix<double, 3, 3, Eigen::RowMajor> >(calibVec.data());
+    Eigen::Matrix3d calibMat
+        = Eigen::Map<Eigen::Matrix<double, 3, 3, Eigen::RowMajor>>(calibVec.data());
     m_pimpl->cameraMatrix = cv::Mat(3, 3, CV_64F);
     cv::eigen2cv(calibMat, m_pimpl->cameraMatrix);
 
     Eigen::Matrix<double, 5, 1> distCoeffVec;
     if (!handle->getParameter("distortion_coefficients", distCoeffVec))
     {
-        log()->error("{} "
-        "The parameter handler could not find \" distortion_coefficients \" in the configuration file.",
-        printPrefix);
+        log()->error("{} The parameter handler could not find \" distortion_coefficients \" in the "
+                     "configuration file.",
+                     printPrefix);
         return false;
     }
 
@@ -151,12 +157,12 @@ bool ArucoDetector::initialize(std::weak_ptr<IParametersHandler> handler)
 
 bool ArucoDetector::setImage(const cv::Mat& inputImg, double timeNow)
 {
-    std::string_view printPrefix = "[ArucoDetector::setImage] ";
+    constexpr auto printPrefix = "[ArucoDetector::setImage]";
     if (!m_pimpl->initialized)
     {
-        log()->error("{} "
-        "Unable to set image. Please initialize the ArucoDetector before setting the image.",
-        printPrefix);
+        log()->error("{} Unable to set image. Please initialize the ArucoDetector before setting "
+                     "the image.",
+                     printPrefix);
         return false;
     }
 
@@ -168,25 +174,22 @@ bool ArucoDetector::setImage(const cv::Mat& inputImg, double timeNow)
 
 bool ArucoDetector::advance()
 {
-    std::string_view printPrefix = "[ArucoDetector::advance] ";
+    constexpr auto printPrefix = "[ArucoDetector::advance]";
     if (!m_pimpl->initialized)
     {
-        log()->error("{} "
-        "Unable to run advance(). Please initialize the ArucoDetector first.",
-        printPrefix);
+        log()->error("{} Unable to run advance(). Please initialize the ArucoDetector first.",
+                     printPrefix);
         return false;
     }
 
     if (m_pimpl->currentImg.empty())
     {
-        log()->error("{} "
-        "Unable to run advance(). Please set an image first.",
-        printPrefix);
+        log()->error("{} Unable to run advance(). Please set an image first.", printPrefix);
         return false;
     }
 
     m_pimpl->resetBuffers();
-    std::vector<std::vector<cv::Point2f> > detectedmarkerCorners;
+    std::vector<std::vector<cv::Point2f>> detectedmarkerCorners;
     cv::aruco::detectMarkers(m_pimpl->currentImg,
                              m_pimpl->dictionary,
                              m_pimpl->currentDetectedMarkerCorners,
@@ -205,9 +208,9 @@ bool ArucoDetector::advance()
         {
             cv::Rodrigues(m_pimpl->currentDetectedMarkersRotVecs[idx], m_pimpl->R);
             cv::cv2eigen(m_pimpl->R, m_pimpl->Reig);
-            m_pimpl->teig <<  m_pimpl->currentDetectedMarkersTransVecs[idx](0),
-                              m_pimpl->currentDetectedMarkersTransVecs[idx](1),
-                              m_pimpl->currentDetectedMarkersTransVecs[idx](2);
+            m_pimpl->teig << m_pimpl->currentDetectedMarkersTransVecs[idx](0),
+                m_pimpl->currentDetectedMarkersTransVecs[idx](1),
+                m_pimpl->currentDetectedMarkersTransVecs[idx](2);
 
             m_pimpl->poseEig = toEigenPose(m_pimpl->Reig, m_pimpl->teig);
             ArucoMarkerData markerData{m_pimpl->currentDetectedMarkerIds[idx],
@@ -218,16 +221,15 @@ bool ArucoDetector::advance()
         m_pimpl->out.timeNow = m_pimpl->currentTime;
     }
 
-
     return true;
 }
 
-const ArucoDetectorOutput& ArucoDetector::get() const
+const ArucoDetectorOutput& ArucoDetector::getOutput() const
 {
     return m_pimpl->out;
 }
 
-bool ArucoDetector::isValid() const
+bool ArucoDetector::isOutputValid() const
 {
     if (!m_pimpl->initialized)
     {
@@ -248,13 +250,12 @@ bool ArucoDetector::getDetectedMarkerData(const int& id, ArucoMarkerData& marker
     return true;
 }
 
-
 bool ArucoDetector::getImageWithDetectedMarkers(cv::Mat& outputImg,
                                                 const bool& drawFrames,
                                                 const double& axisLengthForDrawing)
 {
     std::size_t nrDetectedMarkers = m_pimpl->currentDetectedMarkerIds.size();
-    if (m_pimpl->currentImg.empty() ||  nrDetectedMarkers <= 0)
+    if (m_pimpl->currentImg.empty() || nrDetectedMarkers <= 0)
     {
         return false;
     }
@@ -281,12 +282,10 @@ bool ArucoDetector::getImageWithDetectedMarkers(cv::Mat& outputImg,
                                 m_pimpl->currentDetectedMarkersTransVecs[idx],
                                 axisLengthForDrawing);
         }
-
     }
 
     return true;
 }
-
 
 void ArucoDetector::Impl::resetBuffers()
 {